requirements.txt
asgiref==3.2.7
Django==2.2.12
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-notifications-hq==1.6.0
jsonfield==3.1.0
mysqlclient==1.4.6
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
